Updated to typescript
