+++ package.json
@@ -2,17 +2,18 @@
   "name": "repo2pdf",
   "version": "1.2.1",
   "description": "A Node.js utility for generating a PDF document from a GitHub repository",
-  "main": "clone.cjs",
+  "main": "dist/clone.js",
   "bin": {
-    "repo2pdf": "./clone.cjs"
+    "repo2pdf": "dist/clone.js"
   },
   "scripts": {
-    "start": "node clone.cjs",
+    "start": "node dist/clone.js",
+    "watch": "tsc -w",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/BankkRoll/Repo-to-PDF.git"
+    "url": "https://github.com/malpou/Repo-to-PDF.git"
   },
   "keywords": [
     "github",
@@ -36,6 +37,8 @@
   "homepage": "https://github.com/BankkRoll/Repo-to-PDF#readme",
   "documentation": "https://github.com/BankkRoll/Repo-to-PDF#readme",
   "dependencies": {
+    "@types/inquirer": "^9.0.3",
+    "@types/pdfkit": "^0.12.10",
     "chalk": "^5.2.0",
     "highlight.js": "^11.8.0",
     "inquirer": "^9.2.6",
@@ -43,7 +46,8 @@
     "ora": "^6.3.1",
     "pdfkit": "^0.13.0",
     "puppeteer": "^20.7.3",
-    "simple-git": "^3.18.0"
+    "simple-git": "^3.18.0",
+    "typescript": "^5.1.3"
   },
   "engines": {
     "node": ">=14.0.0"
